Add hospital data, but put it away for now, also clean things up
