+++ packages/pipelines/src/corona/datasets/getCoronaDataArc.ts
@@ -1,8 +1,8 @@
 import { ICoronaDataPoint } from "@corona/api";
 import fetch from "node-fetch";
-import { getArcgisFipsCode } from "../utils/getCoronaDataScraperFipsCode";
-import { getTotalBreakdowns, ICountiesKeyed, IStatesKeyed, ITotalBreakdown } from "./shared";
-import { logArcgisAnomalies } from "../utils/logAnomalies";
+import { getArcgisFipsCode } from "../../utils/getCoronaDataScraperFipsCode";
+import { getTotalBreakdowns, ICountiesKeyed, IStatesKeyed, ITotalBreakdown } from "../shared";
+import { logArcgisAnomalies } from "../../utils/logAnomalies";
 
 interface IRawArcCoronaData {
     attributes: IArcCoronaData;
